Majore reworking to allow class-based system to function
