+++ pycbc/fft/backend_cpu.py
@@ -15,29 +15,25 @@
 #  Free Software Foundation, Inc., 59 Temple Place, Suite 330, Boston,
 #  MA  02111-1307  USA
 
-import pycbc
-from .core import _list_available, _update_global_available
-from .core import _all_backends_list, _all_backends_dict
-from pycbc.scheme import CUDAScheme
+from .core import _list_available
 
-_backend_dict = {'cuda' : 'cufft',
-                 'pyfft' : 'cuda_pyfft'}
-_backend_list = ['cuda','pyfft']
+_backend_dict = {'fftw' : 'fftw',
+                 'mkl' : 'mkl',
+                 'lal' : 'lalfft',
+                 'numpy' : 'npfft'}
+_backend_list = ['fftw','mkl','lal','numpy']
 
-_alist = []
-_adict = {}
+_alist, _adict = _list_available(_backend_list, _backend_dict)
 
-if pycbc.HAVE_CUDA:
-    _alist, _adict = _list_available(_backend_list,_backend_dict)
+cpu_backend = 'fftw'
 
-CUDAScheme.fft_backends_list = _alist
-CUDAScheme.fft_backends_dict = _adict
+def set_backend(backend_list):
+    global cpu_backend
+    for backend in backend_list:
+        if backend in _alist:
+            cpu_backend = backend
+            break
 
-_update_global_available(_alist,_adict,_all_backends_list,_all_backends_dict)
+def get_backend():
+    return _adict[cpu_backend]
 
-# No need to keep temp variables around
-
-del _alist
-del _adict
-del _backend_list
-del _backend_dict
