+++ pycbc/fft/backend_opencl.py
@@ -16,28 +16,27 @@
 #  MA  02111-1307  USA
 
 import pycbc
-from .core import _list_available, _update_global_available
-from .core import _all_backends_list, _all_backends_dict
-from pycbc.scheme import CUDAScheme
+from .core import _list_available
 
-_backend_dict = {'cuda' : 'cufft',
-                 'pyfft' : 'cuda_pyfft'}
-_backend_list = ['cuda','pyfft']
+_backend_dict = {'pyfft' : 'cl_pyfft'}
+_backend_list = ['pyfft']
 
 _alist = []
 _adict = {}
 
-if pycbc.HAVE_CUDA:
-    _alist, _adict = _list_available(_backend_list,_backend_dict)
+if pycbc.HAVE_OPENCL:
+    _alist, _adict = _list_available(_backend_list, _backend_dict)
+    opencl_backend = 'pyfft'
+else:
+    opencl_backend = None
 
-CUDAScheme.fft_backends_list = _alist
-CUDAScheme.fft_backends_dict = _adict
+def set_backend(backend_list):
+    global opencl_backend
+    for backend in backend_list:
+        if backend in _alist:
+            opencl_backend = backend
+            break
 
-_update_global_available(_alist,_adict,_all_backends_list,_all_backends_dict)
+def get_backend():
+    return _adict[opencl_backend]
 
-# No need to keep temp variables around
-
-del _alist
-del _adict
-del _backend_list
-del _backend_dict
